requirements.txt

langchain==0.0.267
openai==0.28.0.0
streamlit==1.25.0
einops==0.6.1
bitsandbytes==0.41.1
accelerate==0.21.0
pdfminer.six==20221105
bs4==0.0.1
sentencepiece==0.1.99
six==1.16.0
requests==2.31.0
uvicorn==0.18.3
streamlit-chat
python-dotenv
